terraform/cost_calculator/requirements.txt
BeautifulSoup4
python-hcl2
requests
tabulate
